Add GNU debug link to load debug symbols in gdb
Along with the modifications to the strip call, this seems to fix the
broken relationship between the two files. Also, one no longer needs to
load the debug symbols manually.

Ideally, one can put the runtime.debug file in the same directory as the
AppImage, then run gdb my.AppImage.