Improve build docs
This commit eliminates redundancies between README.md and BUILD.md. Not having to maintain redundant documentation is a big improvement as there is no chance they will be out of sync. Plus, the docs in BUILD.md are a bit more sophisticated while still allowing @probonopd to copy-and-paste commands for his workflow.

Additionally, the environment variable used in the chroot scripts has been fixed.

The commit further adds a short introduction explaining the differences between the build scripts and recommending Docker for regular users.